Fix: TotalTransportedWeight seems to be float

+++ block_trailer_utilization_log.go
@@ -5,10 +5,10 @@ func init() {
 }
 
 type TrailerUtilizationLog struct {
-	Entries                 []Ptr `sii:"entries"` // Needs verification
-	TotalDrivenDistanceKM   int64 `sii:"total_driven_distance_km"`
-	TotalTransportedCargoes int64 `sii:"total_transported_cargoes"`
-	TotalTransportedWeight  int64 `sii:"total_transported_weight"`
+	Entries                 []Ptr   `sii:"entries"` // Needs verification
+	TotalDrivenDistanceKM   int64   `sii:"total_driven_distance_km"`
+	TotalTransportedCargoes int64   `sii:"total_transported_cargoes"`
+	TotalTransportedWeight  float32 `sii:"total_transported_weight"`
 
 	blockName string
 }
